@oscardssmith 2023-04-21
is it also slower in the range where the old one was accurate?