@heltonmc 2023-04-24
Yep it is definitely slower. But I think it's a hard comparison because the old version I'm not for sure we could say was "accurate" in any range.

```
# old version
julia> x = 2.4f0
2.4f0
julia> 1 - besselj0(Float64(x)) / besselj0(x)
7.178780837779897e-6

# this version
julia> x = 2.4f0
2.4f0

julia> 1 - besselj0(Float64(x)) / besselj0(x)
-2.4906737694507797e-7
```
Some larger comparisons on accuracy...
```
# old version
julia> x = Float32.(rand(1000)*50);

julia> err = @. 1 - besselj0(x) / besselj0(Float64(x));

julia> maximum(err)
0.00024567309384593283

julia> median(err)
-2.4034297774200297e-8

julia> mean(err)
2.387311999333397e-7

# new version
julia> x = Float32.(rand(1000)*50);

julia> err = @. 1 - besselj0(x) / besselj0(Float64(x));

julia> maximum(err)
2.1666121468921773e-5

julia> median(err)
-1.5063105207602234e-8

julia> mean(err)
-2.003219752944041e-8
```

But ya it is a performance degradation.
```julia
# old
julia> @benchmark besselj0(x) setup=(x=Float32(rand() + 2.0))
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  9.000 ns … 25.500 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     9.083 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.090 ns ±  0.252 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                         █                                    
  ▂▁▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▂ ▂
  9 ns           Histogram: frequency by time        9.21 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# new
julia> @benchmark besselj0(x) setup=(x=Float32(rand() + 2.0))
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  15.364 ns … 37.074 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     15.489 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.500 ns ±  0.417 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                     ▃         █                               
  ▂▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▂▃▁▁▁▁▁▁▁▁▂ ▂
  15.4 ns         Histogram: frequency by time        15.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```