@heltonmc 2023-04-24
Side note: It would be really nice to have some CI setup for benchmarking all these functions. Maybe not have it run on every pull request but more on command using either PkgBenchmark.jl or the recent https://github.com/MilesCranmer/AirspeedVelocity.jl which looked interesting.