@heltonmc 2023-04-30
@oscardssmith How do we feel about merging this? I think I would prefer the accuracy improvements in this and can improve performance over time if needed.